requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7 # Para seguridad de contraseñas
scikit-learn==1.3.2 # Para la IA de recomendaciones
numpy==1.26.2